# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LinguaLeap System resources",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/languages/exercise/{exercise_id}/submit": {
            "post": {
                "description": "Submit an exercise for evaluation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Submit an exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exercise_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit Exercise",
                        "name": "submit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionResponse"
                        }
                    }
                }
            }
        },
        "/languages/exercises/{language_code}": {
            "get": {
                "description": "Get exercises for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get exercises",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionResponse"
                        }
                    }
                }
            }
        },
        "/languages/lesson/{lesson_id}": {
            "get": {
                "description": "Get details for a specific lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get lesson details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/lesson/{lesson_id}/complete": {
            "post": {
                "description": "Mark a lesson as completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Complete a lesson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Complete Lesson",
                        "name": "complete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/lessons/{language_code}": {
            "get": {
                "description": "Get lessons for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get lessons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/start": {
            "post": {
                "description": "Start learning a language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Start learning a language",
                "parameters": [
                    {
                        "description": "Start Learning Language",
                        "name": "start",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.CreateLanguageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.CreateLanguageResponse"
                        }
                    }
                }
            }
        },
        "/languages/vocabulary/{language_code}": {
            "get": {
                "description": "Get vocabulary for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get vocabulary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetVocabularyResponse"
                        }
                    }
                }
            }
        },
        "/languages/vocabulary/{language_code}/add": {
            "post": {
                "description": "Add new vocabulary to a language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Add vocabulary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add Vocabulary",
                        "name": "vocabulary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.AddVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.AddVocabularyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "learn.AddVocabularyRequest": {
            "type": "object",
            "properties": {
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "translation": {
                    "description": "So'z tarjimasi",
                    "type": "string"
                },
                "word": {
                    "description": "So'z",
                    "type": "string"
                }
            }
        },
        "learn.AddVocabularyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Yaratilgan leksika yozuving UUID si",
                    "type": "string"
                }
            }
        },
        "learn.CreateLanguageRequest": {
            "type": "object",
            "properties": {
                "emoji": {
                    "description": "Til flag emojisi",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "name": {
                    "description": "Til nomi, masalan, 'Ingliz tili', 'O'zbek tili'",
                    "type": "string"
                }
            }
        },
        "learn.CreateLanguageResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "Yaratilgan til haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Language"
                        }
                    ]
                }
            }
        },
...
        "learn.GetLessonRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Darsning UUID si",
                    "type": "string"
                }
            }
        },
        "learn.GetLessonResponse": {
            "type": "object",
            "properties": {
                "lesson": {
                    "description": "Dars haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Lesson"
                        }
                    ]
                }
            }
        },
        "learn.GetQuestionRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Savolning UUID si",
                    "type": "string"
                }
            }
        },
        "learn.GetQuestionResponse": {
            "type": "object",
            "properties": {
                "question": {
                    "description": "Savol haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Question"
                        }
                    ]
                }
            }
        },
        "learn.GetVocabularyResponse": {
            "type": "object",
            "properties": {
                "vocabulary": {
                    "description": "Leksika yozuvi haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Vocabulary"
                        }
                    ]
                }
            }
        },
        "learn.Language": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "emoji": {
                    "description": "Til flag emojisi",
                    "type": "string"
                },
                "id": {
                    "description": "Tilning UUID si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "name": {
                    "description": "Til nomi, masalan, 'Ingliz tili', 'O'zbek tili'",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Yangilangan vaqti",
                    "type": "string"
                }
            }
        },
        "learn.Lesson": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "description": "Dars yaratilgan vaqti",
                    "type": "string"
                },
                "id": {
                    "description": "Darsning UUID si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "level": {
                    "description": "Dars darajasi, masalan, 'boshlang'ich', 'o'rta', 'yuqori'",
                    "type": "string"
                },
                "title": {
                    "description": "Dars nomi",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Dars yangilangan vaqti",
                    "type": "string"
                }
            }
        },
        "learn.Question": {
            "type": "object",
            "properties": {
                "correct_answer": {
                    "description": "To'g'ri javob",
                    "type": "string"
                },
                "created_at": {
                    "description": "Yaratilgan vaqti",
                    "type": "string"
                },
                "difficulty": {
                    "description": "Qiyinlik darajasi",
                    "type": "string"
                },
                "explanation": {
                    "description": "Javob izohi",
                    "type": "string"
                },
                "id": {
                    "description": "Savolning UUID si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "options": {
                    "description": "Javob variantlari",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question": {
                    "description": "Savol matni",
                    "type": "string"
                },
                "type": {
                    "description": "Savol turi",
                    "type": "string"
                }
            }
        },
        "learn.Vocabulary": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Yaratilgan vaqti",
                    "type": "string"
                },
                "id": {
                    "description": "Leksika yozuving UUID si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "translation": {
                    "description": "So'z tarjimasi",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Yangilangan vaqti",
                    "type": "string"
                },
                "word": {
                    "description": "So'z",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}